current_requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.1
blinker==1.9.0
brax==0.13.0
cffi==2.0.0
chex==0.1.91
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.1
debugpy==1.8.17
decorator==5.2.1
etils==1.13.0
executing==2.2.1
Farama-Notifications==0.0.4
fasteners==0.20
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
flax==0.11.0
fonttools==4.60.1
fsspec==2025.10.0
glfw==2.10.0
gymnasium==1.2.2
humanize==4.14.0
ImageIO==2.37.2
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.7.0
jaxlib==0.7.0
jaxopt==0.8.5
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.4.0
mujoco-mjx==3.4.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.5
omegaconf==2.3.0
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.30
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
rich==14.2.0
scipy==1.16.3
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tensorboardX==2.6.4
tensorstore==0.1.79
toolz==1.1.0
torch==2.9.1
tornado==6.5.2
traitlets==5.14.3
treescope==0.1.10
trimesh==4.10.1
typing_extensions==4.15.0
wcwidth==0.2.14
Werkzeug==3.1.4
zipp==3.23.0
